Terrain collider problem
I need to figure out how to get the player to detect the terrain collider. I'm thinking of creating objects with a mesh collider that mirrors the terrain abutments they represent. The sun rotation is completed; now, I need to design the timer mechanic, store mechanic, powerup mechanics, UI and UX, and what the Sensei called "plot twists", which consists of the well mechanic. I'll also need to incorporate the eagle and player animations.