Testing, testing, is this on?
I've written the eagle attack logic. At this moment, if the player triggers the eagle by standing directly underneath it, then the eagle will pursue the player by moving in an angle composed of the difference in x position and 20f. If the difference in distance is less than one frame, then the player dies. I also need to code the player attack logic, which requires the staff, along with setting up the UI elements to fit the plot.