Winners and Losers
I worked on endgame animation logic. It mostly works fine, but it says you lose when you win. The idea behind it is that there is an overall animation between two planes, one for winning, one for losing. When the player wins or loses, the parent, which has the animation on it, is activated. One of the two objects is then deactivated, depending on if the player lost or won.