Set adrift on Memory Bliss...
I've created the restart slider and made sure that the high score method works. Next time, I need to expand the high score variable to the beginning scene so that the player can keep track of their progress.

BTW, the score is determined by an aggregate of time remaining, feathers collected, friend_returned, and eagle_blows.