Just cheeky
Unity is being stupid; although I added a new terrain to replace the accidentally deleted one, and the webgl file is finally working completely in unity play, the lighting is sucking cheeks harder than Albert I of Savoy against the Austrians in 1848. I'm trying to fix it, but to no avail. The terrain shader is the issue.