Assertion failed: m_LockCount == 0
It appears that when Collin and I manipulated the WMD shaders in unity to conform to URP, we may have messed up something in the file. Next time, I'm going to reintroduce the WMD shader into the game in hopes that the game will snap back to normal, just like before.