Preprocessor shenanigans
After capping off the endgame system, I attempted to clear the clipping issue, but to no avail. Also, when I tried to build a new WebGL file, a new error appeared which claimed that there was a preprocessor misplaced during the build. I have no idea how to fix it.